feat(entity): add comment, favorite, and like modules with relations
Added new modules for comment, favorite, and like functionalities within the entity
layer, establishing their relationships with the post and user entities.

+++ entity/src/vo/ums_user.rs
@@ -17,7 +17,7 @@ pub struct LoginReq {
 #[derive(Deserialize, Serialize)]
 pub struct LoginRes {
     pub token: String,
-    pub uid: i64,
+    pub uid: i32,
     pub nickname: String,
     pub avatar: String,
 }
